minor fixes in inscribe and unsuscribe

+++ app/Http/Controllers/EventController.php
@@ -98,7 +98,7 @@ class EventController extends Controller
             }
         }
         $user->event()->attach($event);
-        $event->sub_people = $user->sub_people +1;
+        $event->sub_people += 1;
         $event->save();
 
         return redirect()->route('home');
@@ -109,7 +109,15 @@ class EventController extends Controller
         $user = User::find(Auth::id());
         $event = Event::find($id);
 
-        $user->event()->detach($event);
+        $user->event();
+        foreach ($user->event as $eventInscribed){
+            if($eventInscribed->pivot->event_id == $id){
+                $user->event()->detach($event);
+                $event->sub_people -= 1;
+                $event->save();
+                return redirect()->route('home');
+            }
+        }
         return redirect()->route('home');
     }
     public function eventsSubscribe()
